server/requirements.txt
fastapi
uvicorn
playwright
twilio
anthropic
python-dotenv
httpx
supabase
aiofiles
